apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
prisma==0.15.0
reportlab==4.2.5
qrcode[pil]==8.0
pillow==11.1.0
python-dotenv==1.0.1
email-validator>=2.0.0
python-dateutil==2.9.0.post0
slowapi==0.1.9
bleach==6.1.0
psutil>=5.9.0
sentry-sdk[fastapi]


